implementet ability to alter style option

+++ src/main/java/FunctionLayer/entities/StyleOption.java
@@ -47,7 +47,7 @@ public class StyleOption {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
